Simplify ostringstream object instantiation
This update will instantiate stringstream objects in the Constructor
and add them into the buffer_list dynamically, making the class-wide
declarations obsolete and instead relying on the buffer_list as the
central point of access for all buffers.

Is this good or bad? You tell me. But it's at least simpler than it
was previously, and I make sure to avoid memory leaks.