Avert segfault, refactor Scrollable widgets
This update will replace the "list of Widgets" owned by a Scroll Widget with
ONE Scrollable object whose display is being regulated.

It will refactors some of the subclasses of Widget: Scrollable widgets may have
a Scroll button, and so far, only List and Text Widgets are Scrollable.

In addition to modifying the Scroll class itself in order to point to its
Scrollable Widget, it will also update Scroll::move() to check the items before
changing the widget's area_offset.

In effect, this should avert the segmentation fault -- but some more testing
would be ideal.

Notice that the src/CMakeLists.txt has been modified to accommodate the new
class file.